requirements.txt
# Core dependencies
streamlit
ultralytics>=8.0.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Progress Bars and Rich Display
tqdm>=4.65.0
rich>=13.0.0

# Optional - for video processing
python-ffmpeg>=1.0.0
av>=10.0.0

# Optional - for export features
matplotlib>=3.7.0
seaborn>=0.12.0

# System and File Management
python-dotenv>=1.0.0
pyYAML>=6.0.1

# Optional - for development
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0t
flask
